fefactor: extract context, Toolbar, NoOpenScripts components
